services/python/front/pyproject.toml
[tool.poetry]
name = "snet-front"
version = "0.1.0"
packages = [{include = "app"}]
description = "SNet Python front impl"
authors = ["VladNF <j@VladNF.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
environs = "^9.5.0"
asyncpg = "^0.27.0"
pyjwt = "^2.6.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.6.1"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^24.2.0"
psycopg2-binary = "^2.9.5"
pytest-asyncio = "^0.23.5"
pytest-env = "^1.1.3"
httpx = "^0.27.0"
yoyo-migrations = "^8.2.0"
setuptools = "^69.1.1"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.pytest_env]
POSTGRES_PORT = 5532
JWT_PRIVATE_KEY=424342
JWT_PUBLIC_KEY=424342
